chore: fix all linting errors in backend and tests
Resolved 39+ issues including:
- Complexity and line length in holdings.py
- Exception naming and timezone usage in tests
- Import order and type-checking guards
- Subprocess safety in frontend tests

+++ tests/normalization/test_pipeline.py
@@ -25,12 +25,12 @@ if TYPE_CHECKING:
 
 def _shares_raw(value: float) -> int:
     """Convert share counts to the stored 10^-8 representation."""
-    return int(round(value * 10**8))
+    return round(value * 10**8)
 
 
 def _price_raw(value: float) -> int:
     """Convert prices to the stored 10^-8 representation."""
-    return int(round(value * 10**8))
+    return round(value * 10**8)
 
 
 def _purchase_total(value: float) -> float:
